Revert "Alerting: Show annotations markers in TimeSeries panel when using Loki as … (#72084)"
This reverts commit de6ef53c8a317745a874ef6ce8dbb736369fb0af.
