+++ public/app/features/alerting/unified/hooks/useStateHistoryModal.tsx
@@ -2,15 +2,14 @@ import { css } from '@emotion/css';
 import React, { lazy, Suspense, useCallback, useMemo, useState } from 'react';
 
 import { GrafanaTheme2 } from '@grafana/data';
+import { config } from '@grafana/runtime';
 import { Modal, useStyles2 } from '@grafana/ui';
 import { RulerGrafanaRuleDTO } from 'app/types/unified-alerting-dto';
-
-import { getHistoryImplementation } from '../components/rules/state-history/common';
 
 const AnnotationsStateHistory = lazy(() => import('../components/rules/state-history/StateHistory'));
 const LokiStateHistory = lazy(() => import('../components/rules/state-history/LokiStateHistory'));
 
-export enum StateHistoryImplementation {
+enum StateHistoryImplementation {
   Loki = 'loki',
   Annotations = 'annotations',
 }
@@ -21,7 +20,18 @@ function useStateHistoryModal() {
 
   const styles = useStyles2(getStyles);
 
-  const implementation = getHistoryImplementation();
+  // can be "loki", "multiple" or "annotations"
+  const stateHistoryBackend = config.unifiedAlerting.alertStateHistoryBackend;
+  // can be "loki" or "annotations"
+  const stateHistoryPrimary = config.unifiedAlerting.alertStateHistoryPrimary;
+
+  // if "loki" is either the backend or the primary, show the new state history implementation
+  const usingNewAlertStateHistory = [stateHistoryBackend, stateHistoryPrimary].some(
+    (implementation) => implementation === StateHistoryImplementation.Loki
+  );
+  const implementation = usingNewAlertStateHistory
+    ? StateHistoryImplementation.Loki
+    : StateHistoryImplementation.Annotations;
 
   const dismissModal = useCallback(() => {
     setRule(undefined);
