+++ public/app/features/query/state/PanelQueryRunner.ts
@@ -3,7 +3,6 @@ import { Observable, of, ReplaySubject, Unsubscribable } from 'rxjs';
 import { map, mergeMap, catchError } from 'rxjs/operators';
 
 import {
-  ApplyFieldOverrideOptions,
   applyFieldOverrides,
   compareArrayValues,
   compareDataFrameStructures,
@@ -20,18 +19,18 @@ import {
   getDefaultTimeRange,
   LoadingState,
   PanelData,
-  preProcessPanelData,
   rangeUtil,
   ScopedVars,
-  StreamingDataFrame,
   TimeRange,
   TimeZone,
   toDataFrame,
   transformDataFrame,
+  preProcessPanelData,
+  ApplyFieldOverrideOptions,
+  StreamingDataFrame,
 } from '@grafana/data';
 import { getTemplateSrv, toDataQueryError } from '@grafana/runtime';
 import { ExpressionDatasourceRef } from '@grafana/runtime/src/utils/DataSourceWithBackend';
-import { updatePanelDataWithASHFromLoki } from 'app/features/alerting/unified/components/rules/state-history/common';
 import { isStreamingDataFrame } from 'app/features/live/data/utils';
 import { getDatasourceSrv } from 'app/features/plugins/datasource_srv';
 
@@ -331,11 +330,8 @@ export class PanelQueryRunner {
     }
 
     this.subscription = panelData.subscribe({
-      next: async (data) => {
-        this.lastResult = skipPreProcess
-          ? data
-          : await updatePanelDataWithASHFromLoki(preProcessPanelData(data, this.lastResult));
-
+      next: (data) => {
+        this.lastResult = skipPreProcess ? data : preProcessPanelData(data, this.lastResult);
         // Store preprocessed query results for applying overrides later on in the pipeline
         this.subject.next(this.lastResult);
       },
